#individual-assessment: last reading attempt.

+++ individual-assessment/citizen/src/main/java/com/smart_cities/citizen/contracts/GeneratesReadings.java
@@ -3,5 +3,5 @@ package com.smart_cities.citizen.contracts;
 public interface GeneratesReadings {
     public Long getId();
     public String getType();
-    public int generateReading();
+    public Long generateReading();
 }
